Create the initial bibliography & add article
This commit, adds the bibliography database and adds the first article
to it. The second article's URL appears invalid, but if I can find an
open-source version of it, I'll add that one as well.

If you have your PDF viewer set up, and JabRef installed, you should be
able to just open the bibliography.bib file in JabRef and click the PDF
icon on any paper to view it.